voice control car
This project builds a voice controlled car that can be controlled by voice commands which reacts in accordance to the corresponding voice command. Simple voice commands like DANCE , left, right, forward, back, stop, red light , right side yellow and so on.. are used to run the car. These commands are given to Bluetooth module via to android application. The Bluetooth module and control unit are combined to store and test the voice commands. When an instruction for the automobile (car) is identified, a command message is sent to Arduino UNO, This command is analyzed by the microcontroller of arudino and followed up, In the car, Image processing can be utilized to become aware of the shade and the obstacles. This work has been limited to the ZigBee system in the short-range (100mts range), and is linked to the car over long distance via long-range module